apis/requirements.txt
attrs==23.2.0
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
flasgger==0.9.7.1
Flask==3.0.2
flask-swagger-ui==4.11.1
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
load-dotenv==0.1.0
MarkupSafe==2.1.5
mistune==3.0.2
packaging==23.2
pymongo==4.6.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
Werkzeug==3.0.1
zipp==3.17.0
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.23